+++ src/css/App.css
@@ -1,0 +1,7 @@
+.App {
+  height: 400px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
